refactor: organize folder structure
