+++ app/routes/actions/logout.tsx
